+++ core/platform/auth/firebase_auth_services_test.py
@@ -28,7 +28,7 @@ from core.tests import test_utils
 import firebase_admin
 import webapp2
 
-(user_models,) = models.Registry.import_models([models.NAMES.user])
+(auth_models,) = models.Registry.import_models([models.NAMES.auth])
 
 
 class FirebaseAuthServicesPublicApiTest(test_utils.TestBase):
@@ -76,7 +76,7 @@ class FirebaseAuthServicesPublicApiTest(test_utils.TestBase):
     def test_user_id_association_model_is_for_firebase(self):
         self.assertIs(
             firebase_auth_services.UserIdBySubjectIdModel,
-            user_models.UserIdByFirebaseSubjectIdModel)
+            auth_models.UserIdByFirebaseSubjectIdModel)
 
     def test_error_in_initialize_propogates(self):
         app_that_will_not_initialize = self.swap_to_always_raise(
